requirements.txt
Django==1.10.6
mysqlclient==1.3.10
djangorestframework==3.6.2
